[Feat] 폴더링 모듈화
