+++ src/components/folder/PublicMain.jsx
@@ -1,0 +1,9 @@
+export default function PublicMain() {
+    return (
+      <main className="ml-[200px] mt-[80px] flex-1 p-8">
+        <h1 className="text-2xl font-semibold text-[#5f360a]">Public 영역</h1>
+        {/* TODO: fetchPublicFolder 등 붙이면 바로 확장 가능 */}
+      </main>
+    );
+  }
+
