--- src/pages/Private.module.css
@@ -1,28 +1,0 @@
-.container {
-    margin-top: 0px;
-    margin-left: 200px;
-    min-height: 100vh;
-    background-color: #F9F4ED;
-    font-family: 'Noto Sans KR', sans-serif;
-    display: flex;
-    flex-direction: column;
-    width: calc(100% - 200px);
-}
-
-.main {
-    padding: 2rem;
-    display: flex;
-}
-
-.grid {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1.5rem;
-    /* ✅ 카드 사이 간격 늘리기 */
-    width: 100%;
-    max-width: 1000px;
-    /* ✅ 너무 넓게 퍼지는 것 방지 */
-    grid-template-columns: repeat(auto-fill, minmax(100px, 1fr));
-    align-items: start;
-    /* ✅ 세로 정렬 */
-}
